+++ src/UI/Pages/ServiceInventory/Presenters/InstanceActionPresenter.ts
@@ -1,6 +1,6 @@
 import { ReactElement } from "react";
 import { InstanceActions } from "@/UI/Pages/ServiceInventory/Components";
-import { ActionPresenter, ServiceInstanceForAction } from "@/UI/Contracts";
+import { ActionPresenter, ServiceInstanceForAction } from "@/UI/Presenters";
 import { ServiceModel } from "@/Core";
 
 export class InstanceActionPresenter implements ActionPresenter {
