+++ src/UI/Pages/ServiceInventory/Presenters/InventoryTablePresenter.ts
@@ -1,7 +1,11 @@
 import { ReactElement } from "react";
 import { Id, Row, ServiceInstanceModelWithTargetStates } from "@/Core";
 import { words } from "@/UI/words";
-import { ActionPresenter, TablePresenter, DatePresenter } from "@/UI/Contracts";
+import {
+  ActionPresenter,
+  TablePresenter,
+  DatePresenter,
+} from "@/UI/Presenters";
 import { AttributesPresenter } from "./AttributesPresenter";
 import { StatePresenter } from "./StatePresenter";
